Merging philippkeller/hexo-theme-ewal.git into here so I don't need to keep both repositories in sync.
I hope I'm not adding hassle to anyone with that. I inherited this repo
as the maintainer before did not use hexo any more and did not want to
support this any further. This commit is quite a shift in style and also
probably in the sets of supported widgets.

That said, it still is based on twitter bootstrap and supports
highlightjs, which are IMO the two main features of this repo.